feat : 클래스 분리
- 로직, dto 분리
